Fix data type in album entity

+++ src/main/java/de/htw/webtech/web/Album.java
@@ -1,9 +1,6 @@
 package de.htw.webtech.web;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
 import java.util.Set;
 
 @Entity
@@ -11,7 +8,7 @@ public class Album {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false)
     private String name;
@@ -24,11 +21,11 @@ public class Album {
     }
     public Album() {}
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
